Fix for missing unregister cmd in wifi_rtt_range_cancel
Issue:
Continuous "WifiHAL : Failed to add command 253272: 0xb400007edaf95de0 at 64, reached max limit 64" error has occurred while running the multiple AP RTT

Analysis:
When vendor hal is sent with the range_request, each request is registered as cmd and there is a max limit to register the cmd which is 64.
In this error case, because of below 2 reasons, register_cmd max limit can be hit.
1. Missing the corresponding RTT event, upon which cmd is unregistered.
2. Missing the cancel range request, as part of the range_cancel request, unregistering the cmd need to be there, which is missing.

Need to check why rtt event event is getting missed and need to make sure of the canceling the cmd when range_cancel is receieved.

Fix:
Adding the missing unregister_cmd call and more logs.

Bug: 250396851
Test: RTT comms test done. No regressions observed.

Signed-off-by: Dennis Jeon <dennis.jeon@broadcom.com>
Change-Id: I049d7d00151c0b3fbf0de721cd18d8185f1bfddc